Integrando front

+++ src/main/kotlin/ar/edu/unq/epers/tactics/modelo/habilidades/AtaqueMagico.kt
@@ -3,7 +3,7 @@ package ar.edu.unq.epers.tactics.modelo.habilidades
 import ar.edu.unq.epers.tactics.modelo.Aventurero
 import ar.edu.unq.epers.tactics.modelo.dado.Dado
 
-class AtaqueMagico(val poderMagicoEmisor: Int, val nivelEmisor: Int, aventureroReceptor: Aventurero, val dado: Dado) : Habilidad(aventureroReceptor) {
+class AtaqueMagico(val poderMagicoEmisor: Double, val nivelEmisor: Int, aventureroReceptor: Aventurero, val dado: Dado) : Habilidad(aventureroReceptor) {
 
     companion object {
         fun para(aventureroEmisor: Aventurero, aventureroReceptor: Aventurero, dado: Dado): AtaqueMagico {
